requirements.txt
streamlit
openai
pypdf
faiss-cpu
numpy
